pyproject.toml
[project]
name = "langgraph-chatbot"
version = "0.1.0"
description = "AI Chatbot with LangGraph, multi-tool support, and human-in-the-loop"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "streamlit>=1.39.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "requests>=2.31.0",
    "tavily-python>=0.3.0",
    "reportlab>=4.0.0",
    "markdown>=3.5.0",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "langgraph-checkpoint-sqlite>=3.0.2",
]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
